Faster bit processing
Profiling revealed that the 3rd party bitstring package the reason
that CPTV parsing was so slow. A BitStream class has now been
implemented which is much faster. The API is friendlier to use as
well.

A 12s CPTV file previously took about 12.6s to parse on my machine and
is now down to under 2.5s (~80% faster).

BitStream has py.test based unit tests.